feat: implement lock when deleting author if owning books

+++ src/services/autor.service.js
@@ -1,4 +1,5 @@
 import AutorRepository from '../repositories/autor.repository.js';
+import LivroRepository from '../repositories/livro.repository.js';
 
 async function createAutor(autor) {
   return await AutorRepository.insertAutor(autor);
@@ -9,6 +10,10 @@ async function updateAutor(autor) {
 }
 
 async function deleteAutor(id) {
+  const livros = await LivroRepository.getLivrosPorAutor(id);
+  if (livros) {
+    throw new Error('O autor possui livros');
+  }
   await AutorRepository.deleteAutor(id);
 }
